requirements.txt
Flask
pandas
numpy
requests
faiss-cpu
sentence-transformers